requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
mistralai==1.7.1
pydantic==2.11.5
pydantic-settings==2.1.0
python-multipart==0.0.9
pytest==8.0.1
pytest-asyncio==0.23.5 